refactor evaluation metrics for translation task
